django/requirements.txt
Django==1.9.4
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-filter==0.13.0
django-grappelli==2.8.1
djangorestframework==3.3.2
Markdown==2.6.5
MySQL-python==1.2.5
sqlparse==0.1.19
